flake.toml - first_steps_with_nix_v2.md
In round one focus was on the `nix` command more then anything else.
While we already expected, the need to write `flake.nix` and the need to
know Nix language very early on almost destroys the nice user
experience.

In this round we tried to explore the idea:

> Changing to some known format (toml, json, yaml, ...) how far could we
> use Nix and still not need to know about Nix language.

The reason to explore this idea was the realization that learning Nix
language is quite a big barrier. Currently we require to know Nix
language at quite an early stage - right after using `nix-shell -p`. How
far could we allow new user of Nix to use Nix without knowing Nix and
using a simpler and more familiar format instead. Clearly this is not
going to replace Nix, but only to help on-board new Nix users, until they
reach a problem where learning Nix language is worthwhile of their
investment. And obligatory finish of a paragraph with a cliché quote:

> Simple things should be simple and complex things should be possible.